+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added `Strava::Api::Client#activity_photos` - [@dblock](https://github.com/dblock).
 * [#2](https://github.com/dblock/strava-ruby-client/issues/2): Added `Strava::Api::Client#deauthorize` - [@dblock](https://github.com/dblock).
+* [#15](https://github.com/dblock/strava-ruby-client/issues/15): Added `strava-oauth-token` and `strava-webhooks` to gem executables - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.3.0 (2018/12/03)